implemented checking of user to authorize updates and deletions of only their own posts

+++ app/routers/post.py
@@ -12,15 +12,17 @@ router = APIRouter(
 )
 
 @router.get("/", response_model=list[PostRead])
-def get_posts(session: SessionDep, current_user: int = Depends(tm.get_current_user)):
+def get_posts(session: SessionDep, current_user = Depends(tm.get_current_user)):
     posts = session.exec(select(Post)).all()
 
     return posts
 
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=PostRead)
-def create_post(post: PostCreate, session: SessionDep, current_user: int = Depends(tm.get_current_user)):
-    #print(current_user.email)
+def create_post(post: PostCreate, session: SessionDep, current_user = Depends(tm.get_current_user)):
+    
+    post.owner_id = current_user.id
     new_post = Post.model_validate(post)
+    
     session.add(new_post)
     session.commit()
     session.refresh(new_post)
@@ -29,7 +31,7 @@ def create_post(post: PostCreate, session: SessionDep, current_user: int = Depen
 
 
 @router.get("/{id}", response_model=PostRead)
-def get_post(id: int, session: SessionDep, current_user: int = Depends(tm.get_current_user)):
+def get_post(id: int, session: SessionDep, current_user= Depends(tm.get_current_user)):
     post = session.get(Post, id)
 
     if not post:
@@ -38,27 +40,34 @@ def get_post(id: int, session: SessionDep, current_user: int = Depends(tm.get_cu
     return post
 
 @router.delete("/{id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_post(id: int, session: SessionDep, current_user: int = Depends(tm.get_current_user)):
+def delete_post(id: int, session: SessionDep, current_user = Depends(tm.get_current_user)):
+
     post = session.get(Post, id)
 
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"post with id: {id} was not found")
-    
-    session.delete(post)
-    session.commit()
 
-    return Response(status_code=status.HTTP_204_NO_CONTENT)
+    if current_user.id == post.owner_id:
+        session.delete(post)
+        session.commit()
+
+        return Response(status_code=status.HTTP_204_NO_CONTENT)
+    else:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
 
 @router.put("/{id}", response_model=PostRead, status_code=status.HTTP_200_OK)
-def update_post(id: int, post: PostUpdate, session: SessionDep, current_user: int = Depends(tm.get_current_user)):
+def update_post(id: int, post: PostUpdate, session: SessionDep, current_user = Depends(tm.get_current_user)):
     db_post = session.get(Post, id)
 
     if not db_post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"post with id: {id} was not found")
     
-    db_post.sqlmodel_update(post.model_dump(exclude_unset=True))
-    session.add(db_post)
-    session.commit()
-    session.refresh(db_post)
+    if current_user.id == db_post.owner_id:
+        db_post.sqlmodel_update(post.model_dump(exclude_unset=True))
+        session.add(db_post)
+        session.commit()
+        session.refresh(db_post)
 
-    return db_post
+        return db_post
+    else:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
